HTL Jython API
Full working HTL Jython API.
I had to create a .jar file for the procedural API, and I had to
recreate the HTL_API.jar file to match the procedural API's folder and
file structure. This includes an updated Engine folder from the
GameEngine project. To reduce the overall project size, the archived
HTL_API.jar files that were inside of the original HTL_API.jar file
could be removed, but I left them in for consistency.
This push includes all 11 Labs that are present in the procedural
project, rewritten in python. A note; the 6th lab only included a
constructor that had super(). This created a blank project. The python
equivalent is nearly an empty file, so I'm not sure of it's usefulness
for teaching.