SpaceSmasher Level methods
This commit includes the previously unfinished setLevel(), as well as
getLevel(), and nextLevel().
setLevel() now correctly displays the proper level text for the game.
nextLevel() takes the level variable already being saved and reset in
the API and increments it by 1, then sets the level text.
getLevel() returns the aforementioned level variable.

It's important to note that before adding new blocks to the level, the
blockSet should be cleared.

This commit also includes an example project directly derived from the
SpaceSmasherGame project. This new project, SpaceSmasherLevelExample
includes three total levels (0 through 2) each with a varying number of
blocks per level.